Implement concurrent server side request processing using a threadpool
-TODO: ServerLogic might return the User's UUID after successful login
in order for the client to perform the logout without storing the user's
credentials. This would probably complicate the underlying structure
of TcpMessage because it would still need to return a text message for
the client. Otherwise the client should analyze the response and
aknowledge that it is indeed an UUID instead of a regular text message.