Implement end-edit section request
When the client sends an edit request for a section, the server returns
its content which will the be saved in a cached file on the client
machine. The user is supposed to edit the section somehow (with an
external editor) and the commit the changes back to the server with and
end-edit request.